+++ Docker/run-uwsgi.sh
@@ -3,12 +3,11 @@
 export DEBIAN_FRONTEND=noninteractive
 export DA_ROOT="${DA_ROOT:-/usr/share/docassemble}"
 export DA_CONFIG_FILE="${DA_CONFIG:-${DA_ROOT}/config/config.yml}"
-export DAPYTHONVERSION="${DAPYTHONVERSION:-3}"
 export DA_DEFAULT_LOCAL="local3.8"
 
 export DA_ACTIVATE="${DA_PYTHON:-${DA_ROOT}/${DA_DEFAULT_LOCAL}}/bin/activate"
 source "${DA_ACTIVATE}"
-source /dev/stdin < <(python -m docassemble.base.read_config "$DA_CONFIG_FILE")
+source /dev/stdin < <(python -m docassemble.base.read_config --limited "$DA_CONFIG_FILE" )
 
 set -- $LOCALE
 export LANG=$1